agents/stay_agent/requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
groq
httpx
